backend/requirements.txt
Flask==3.0.2
pdfplumber==0.10.3
transformers==4.37.2
torch==2.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1
requests==2.31.0 